// YOUCOM.ReserVook/src/app/features/master/agent/model/agentinfo.model.ts
import { Base } from '../../../../shared/model/baseinfo.model'

export class AgentInfo implements Base {
  companyNo: string;
  agentCode: string;
  agentName: string;
  remarks: string;
  displayOrder: number;
  status: string;
  version :number;
  creator:string;
  updator:string;
  cdt:string;
  udt:string;
}